backend/requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
bleach==3.2.1
colour==0.1.5
Django==3.1.1
django-bleach==0.6.1
django-ckeditor==5.9.0
django-colorfield==0.3.2
django-common-utils==1.6.0
django-filter==2.3.0
django-hint==0.1.5
django-js-asset==1.2.2
django-lifecycle==0.7.7
django-simple-history==2.11.0
djangorestframework==3.11.1
htmlmin==0.1.12
lorem==0.1.1
Markdown==3.2.2
names==0.3.0
packaging==20.4
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urlman==1.3.0
webencodings==0.5.1
